tools/create_mask/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.2
python-doctr[torch]